[base] Move BoundaryElementProvider to base components directory
