+++ packages/@sanity/base/src/__legacy/@sanity/components/index.ts
@@ -1,2 +1,3 @@
+export * from './boundaryElement'
 export * from './hooks'
 export * from './types'
