+++ packages/@sanity/form-builder/src/inputs/PortableText/Input.tsx
@@ -19,8 +19,8 @@ import {uniqueId, isEqual} from 'lodash'
 import ActivateOnFocus from 'part:@sanity/components/utilities/activate-on-focus'
 import {ChangeIndicatorWithProvidedFullPath} from '@sanity/base/lib/change-indicators'
 import {Layer} from 'part:@sanity/components/layer'
+import {BoundaryElementProvider} from '@sanity/base/__legacy/@sanity/components'
 import PatchEvent from '../../PatchEvent'
-import {BoundaryElementProvider} from './boundaryElement'
 import styles from './PortableTextInput.css'
 import {BlockObject} from './Objects/BlockObject'
 import {InlineObject} from './Objects/InlineObject'
