+++ packages/@sanity/form-builder/src/inputs/PortableText/Objects/renderers/PopoverObjectEditing.tsx
@@ -12,9 +12,9 @@ import {
 } from '@sanity/portable-text-editor'
 import {FormFieldPresence, PresenceOverlay} from '@sanity/base/presence'
 import {Path, Marker, SchemaType} from '@sanity/types'
+import {useBoundaryElement} from '@sanity/base/__legacy/@sanity/components'
 import {FormBuilderInput} from '../../../../FormBuilderInput'
 import {PatchEvent} from '../../../../PatchEvent'
-import {useBoundaryElement} from '../../boundaryElement'
 
 interface Props {
   editorPath: Path
